Successfully connected to mongoDB
