+++ src/mongo/models/news.ts
@@ -1,5 +1,4 @@
-import mongoose, { Schema, Document, connect } from 'mongoose'
-import fs from 'fs'
+import mongoose, { Schema, Document } from 'mongoose'
 
 export type News = Document & {
 	title: string,
@@ -12,20 +11,5 @@ const NewsSchema: Schema = new Schema({
 	body: { type: String, required: false, unique: false },
 	tags: [{ type: String, required: false, unique: false }]
 })
-
-
-async function run(): Promise<void> {
-	const pem = fs.readFileSync('./certs/mongoDB-heart-api-X509.pem')
-
-	await connect('mongodb+srv://heart-api-cluster.g0vnw.mongodb.net/news', {
-		useNewUrlParser: true,
-		useUnifiedTopology: true,
-		authMechanism: 'MONGODB-X509',
-		authSource: '$external',
-		sslKey: pem,
-		sslCert: pem,
-		// authSource: './certs/mongoDB-heart-api-X509.pem'
-	})
-}
 
 export default mongoose.model<News>('News', NewsSchema)
